WIP: Parameterize over effect typeo.
The goal of this commit is to allow the execute method of elastic4s to operate
over arbitrary effect types, eg: instead of returning a Scala Future, we can easily return
a Twitter Future, cats IO, or scalaz Task. The benefit of this should be pretty obvious: Manyusers (myself, for instance) don't use Scala Futures, and having to import an execution context and dispatch unnecessary work into a thread pool, while not critically important, is pretty annoying.

The downside is that users will need to include an extra import (to pull in the correct
executor), along with adding extra complexity to the code base.

Notes on the PR:

1. This currently compiles and works. The tests module fails to compile, but I'm fairly confident it will work once the implicits are available. I wanted to get this in front of you before doing all of that work though, because there are a ton of tests!

2. The files you want to focus on are `FromListener`, a typeclass that describes a type that can convert from a `ResponseListener`, along with `HttpClient`, `HttpExecutable` and `ElasticSearchJavaRestClient`, which are changed to operate over the FromListener typeclass. Everything else is just find and replace, for the most part.

3. Ideally, I think we'd parameterize the effect on the client rather than on the async/execute methods, but it wasn't immediately obvious how this can be done without a really aggressive refactor.

4. The only thing that I think really needs a closer look is the SearchScroll api, because it uses Await to wait for each step. I added an `Awaitable` typeclass to deal with this but I dont think it's ideal, and it's totally untested.

5. While I'm confident that this more or less works, it's rough, needs documentation, needs tests, and needs a lot of cleanup. I'm making this PR to see if you're interested in it before I do all the extra work. This surprisingly only took a few hours, so if you're not interested in taking things in this direction, I won't be angry.

Let me know if you have any questions!