Fixes GeoHashGridBucket keys been returned as burnt in Option toString values.
ie. "Some(q1)"

There are two existing patterns to choose from in the agg responses file; map the key to an Option[String] (like DateRangeBucke) or make an optimistic get from that map and return String (like GeoDistanceBucket).
The latter seems to be most consistent with the other geo aggregation bucket types.